+++ examples/softmax.py
@@ -9,9 +9,9 @@ import numpy as np
 from neuronpy.core.language import bfloat16
 
 from autotune.cache.visualize import plot_metric
-from autotune.core.utils import GEMMCompatibility
-from autotune.tune.benchmark import Benchmark
-from autotune.tune.job import ProfileJobs
+from autotune.core.benchmark import Benchmark
+from autotune.core.job import ProfileJobs
+from autotune.modules.matmul import GEMMCompatibility
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from kernel_library.softmax import softmax_gemm_correctness_postprocessing
